requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==3.0.3
numpy==1.26.4
pandas==2.2.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
Faker==24.4.0
